multagent/edge/requirements_edge.txt
# Edge Client Dependencies (minimal, runs on Raspberry Pi / laser car)
opencv-python>=4.8.0
websockets>=12.0
pyserial>=3.5
pyyaml>=6.0
numpy>=1.24.0
